pyproject.toml
[tool.poetry]
name = "lginnotek"
version = "0.1.0"
description = ""
authors = ["Madung2 <tuliphan91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymupdf = "^1.24.5"
streamlit = "^1.35.0"
pypdf2 = "^3.0.1"
python-levenshtein = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
